Rename "typing" to "atyping".
