+++ pylib-v1/ailets/cons/atyping.py
