+++ pylib-v1/ailets/cons/node_runtime.py
@@ -2,7 +2,7 @@ from typing import Dict, Optional, Sequence
 
 from .node_dagops import NodeDagops
 from .streams import Streams
-from .typing import (
+from .atyping import (
     Dependency,
     IEnvironment,
     INodeDagops,
