+++ pylib-v1/ailets/stdlib/messages_to_markdown.py
@@ -2,7 +2,7 @@ import json
 import base64
 import hashlib
 from ailets.cons.typeguards import is_content_item_image, is_content_item_text
-from ailets.cons.typing import (
+from ailets.cons.atyping import (
     Content,
     ContentItemImage,
     INodeRuntime,
