+++ pylib-v1/ailets/tools/get_user_name/call.py
@@ -1,5 +1,5 @@
 import os
-from ailets.cons.typing import INodeRuntime
+from ailets.cons.atyping import INodeRuntime
 from ailets.cons.util import write_all
